nice files arrangements of stocks
