+++ routes/web/SupplyChain/SupplyChain.php
@@ -6,12 +6,4 @@ Route::group(['namespace' => 'SupplyChain', 'middleware' => ['web', 'auth'], 'pr
 
 
 });
-Route::group(['namespace' => 'SupplyChain', 'middleware' => ['web', 'auth'], 'prefix' => 'stock-unit', 'as' => 'unit.'], function () {
-    Route::get('', 'GoodsController@stockUnit')->name('unit');
-    Route::post('store', 'GoodsController@storeUnit')->name('store');
-    Route::get('edit', 'GoodsController@show')->name('edit');
 
-    Route::group(['prefix' => 'datatables', 'as' => 'datatable.'], function () {
-        Route::get('all', 'GoodsController@alldatatable')->name('all');
-    });
-});
